Corrigir erros

+++ Semana02/ValoresEspeciais/Program.cs
@@ -9,7 +9,7 @@ namespace ValoresEspeciais
             Console.WriteLine(int.MaxValue);
             Console.WriteLine(double.MinValue);
             Console.WriteLine(float.PositiveInfinity);
-            Console.WriteLine(short.NegativeInfinity);
+            Console.WriteLine(double.NegativeInfinity);
             Console.WriteLine(double.NaN);    
 
         }
